+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Upcoming Version / (WIP)
 Improvements:
 * Stabilized consul registration and health checks
 * [ODRC-24](https://openlmis.atlassian.net/browse/ODRC-24) Global header and translations implemented for reports
+* [SELV3-748](https://openlmis.atlassian.net/browse/SELV3-748) Improved 'no permission' warning message
 
 5.3.0 / 2025-11-27
 ==================